+++ src/data/crypto-questions.ts
@@ -1,0 +1,63 @@
+export const cryptoQuestions = [
+	{
+		question: 'What is Blockchain?',
+		answer:
+			'Blockchain is like a digital ledger where transactions are recorded in a secure, transparent, and unchangeable way. Imagine it as a chain of blocks, each holding a list of transactions.',
+		emoji: '🔗',
+	},
+	{
+		question: 'What is Cryptocurrency?',
+		answer:
+			'Cryptocurrency is a type of digital or virtual money. Unlike traditional money, it is not physical and is secured using complex computer techniques called cryptography.',
+		emoji: '💰',
+	},
+	{
+		question: 'Why use Cryptocurrencies?',
+		answer:
+			'Some people like the idea of a currency that isn’t controlled by governments. Others appreciate the anonymity in transactions. It can also be an investment or used for online transactions where cryptocurrencies are accepted.',
+		emoji: '🌍',
+	},
+	{
+		question: 'How do I buy Cryptocurrency?',
+		answer:
+			'You can buy cryptocurrencies using online platforms called exchanges. Once purchased, you can store it in a digital wallet.',
+		emoji: '💸',
+	},
+	{
+		question: 'Why are Cryptocurrencies volatile?',
+		answer:
+			'Their value can change rapidly because of factors like news, regulations, adoption rates, and market speculation.',
+		emoji: '📈',
+	},
+	{
+		question: 'Is Cryptocurrency safe?',
+		answer:
+			'Cryptocurrencies use strong cryptographic techniques for security. However, their safety can depend on how they’re used. While transactions are secure, investments can be risky due to market volatility. Additionally, one must be cautious of scams, secure their digital wallets, and use reputable exchanges.',
+		emoji: '🔒',
+	},
+	{
+		question: 'What is Bitcoin?',
+		answer:
+			'Bitcoin is the first and most well-known cryptocurrency. It’s a decentralised digital currency, meaning no central authority controls it, and transactions are directly between users.',
+		emoji: '🪙',
+	},
+	{
+		question: 'What is a Digital Wallet?',
+		answer:
+			'A digital wallet is like a virtual bank account. It allows you to send or receive cryptocurrencies and monitor your balance.',
+		emoji: '💼',
+	},
+
+	{
+		question: 'What is Mining?',
+		answer:
+			'Mining is a process where powerful computers solve complex problems. When they solve it, they are rewarded with new cryptocurrency coins. It’s also how new transactions are added to a blockchain.',
+		emoji: '⛏️',
+	},
+	{
+		question: 'What is an ICO?',
+		answer:
+			'ICO stands for Initial Coin Offering. It’s a way for new cryptocurrencies to raise funding. Think of it as a mix between a crowdfunding campaign and a stock market IPO.',
+		emoji: '🚀',
+	},
+];